Refactored API. Moved logic from endpoints to separate files. Created decorator for Docker client, to make sure client always getting closed after usage.
